+++ pixelperfectengine/src/pixelperfectengine/concrete/elements/listview.d
@@ -702,8 +702,6 @@ public class ListView : WindowElement, ElementContainer, TextInputListener {
 		if (vertSlider) {
 			const Box p = vertSlider.getPosition();
 			if (p.isBetween(mce.x, mce.y)) {
-				mce.x -= p.left;
-				mce.y -= p.top;
 				vertSlider.passMCE(mec, mce);
 				return;
 			}
@@ -711,8 +709,6 @@ public class ListView : WindowElement, ElementContainer, TextInputListener {
 		if (horizSlider) {
 			const Box p = horizSlider.getPosition();
 			if (p.isBetween(mce.x, mce.y)) {
-				mce.x -= p.left;
-				mce.y -= p.top;
 				horizSlider.passMCE(mec, mce);
 				return;
 			}